Added some basic unit things

+++ src/graphics.h
@@ -1,8 +1,8 @@
 #ifndef GRAPHICS_H_INCLUDED
 #define GRAPHICS_H_INCLUDED
 #include <SDL2/SDL.h>
-//#include <SDL2/SDL_opengl.h>
-#include <SDL2/SDL_opengles.h>
+#include <SDL2/SDL_opengl.h>
+//#include <SDL2/SDL_opengles.h>
 #include <string>
 
 #define SCREEN_WIDTH 640
@@ -14,8 +14,8 @@ namespace Graphics
 {
     void init();
     int loadImage(const std::string& filename);
-    void drawSprite(int x, int y, int width, int height, GLuint texture, bool withOffset = true);
-    void drawText(int x, int y, double width, double height, GLuint texture, const std::string& text);
+    void drawSprite(float x, float y, float width, float height, GLuint texture, bool withOffset = true);
+    void drawText(float x, float y, float width, float height, GLuint texture, const std::string& text);
     void refresh();
 }
